HBX-1944: Move class 'org.hibernate.tool.internal.reveng.JdbcCollectionSecondPass' to 'org.hibernate.tool.internal.reveng.binder.CollectionSecondPass'
